modify workflow.sh

+++ ADNI/workflow.sh
@@ -1,32 +1,30 @@
+#!/bin/bash
+# 去重
 cd ./
-for dir in $(ls -d */); do
-  cd $dir
+for dir in */; do
+  cd "$dir"
   subdirs=(*/)
   for ((i=1; i<${#subdirs[@]}; i++)); do
-    rm -r ${subdirs[$i]}
+    rm -r "${subdirs[$i]}"
   done
   cd ..
 done
 echo "去重完毕!"
 
+# 提取
 mkdir -p initdata
 for dir in */; do
   dir_name=${dir%/}
-  if [ "$dir_name" == "initdata" ]; then
-    continue
-  fi
+  [[ "$dir_name" == "initdata" ]] && continue
   last_nii_file=$(find "$dir" -name "*.nii" | tail -1)
-  if [[ -n $last_nii_file ]]; then
-    mv "$last_nii_file" "initdata/${dir_name}.nii"
-  fi
+  [[ -n $last_nii_file ]] && mv "$last_nii_file" "initdata/${dir_name}.nii"
 done
 for dir in */; do
-  if [ "$dir" != "initdata/" ]; then
-    rm -rf "$dir"
-  fi
+  [[ "$dir" != "initdata/" ]] && rm -rf "$dir"
 done
 echo "提取完成!"
 
+# 去除颅骨
 echo "--------------------------------------------------------------"
 echo "开始去除颅骨"
 input_dir="initdata"
@@ -40,6 +38,8 @@ do
   echo "$file 颅骨去除完成"
 done
 echo "颅骨去除完毕!"
+
+# 配准
 echo "--------------------------------------------------------------"
 echo "开始配准"
 template="template152"
@@ -53,6 +53,8 @@ do
   echo "$file 配准完成"
 done
 echo "配准完毕!"
+
+# 平滑
 echo "--------------------------------------------------------------"
 echo "开始平滑"
 gauss="Gauss_smooth"
@@ -65,11 +67,15 @@ do
   echo "$file 平滑完成"
 done
 echo "平滑完毕!"
+
+# 灰度归一化
 echo "--------------------------------------------------------------"
 echo "开始灰度归一化"
 /home/lz/anaconda3/envs/gdesign/bin/python ./normalized.py
 echo "灰度归一化完毕!"
 
-
-
+# 切片
+echo "--------------------------------------------------------------"
+echo "开始切片"
+/home/lz/anaconda3/envs/gdesign/bin/python ./slice3.py
